requirements.txt
autogen-agentchat
autogen-ext[openai]
google-generativeai
dotenv